requirements.txt
# Dependencias unificadas para o pipeline (GitHub Actions / CI)
# Nao inclui streamlit, matplotlib, seaborn, jupyter (apenas para uso local)

pandas>=2.0.0
requests>=2.31.0
tqdm>=4.65.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
